requirements.txt
PySide6==6.2.0
Pillow==8.4.0
